routes: migrate sessions-content + sessions-proposals (~24 routes)
Move to:
- src/server/routes/sessions-content.ts (13): output, mark-read, generate-title,
  title, tool-content, transcript, abort, prompt-sections, draft GET/PUT/POST/DELETE
- src/server/routes/sessions-proposals.ts (6): proposals list,
  proposal/:type GET/DELETE/edit/seed/restore

Drop now-unused imports (getPromptSections, loadPersistedPromptSections,
sessionFileRead, readTranscript, TranscriptReaderError, proposal-files
imports).

Net: server.ts -433 lines (3433 → 3000). No behaviour change.
