+++ src/server/server.ts
@@ -23,7 +23,7 @@ import { RoleStore } from "./agent/role-store.js";
 import { RoleManager } from "./agent/role-manager.js";
 import { ToolManager } from "./agent/tool-manager.js";
 
-import { getPromptSections, initPromptDirs, loadPersistedPromptSections } from "./agent/system-prompt.js";
+import { initPromptDirs } from "./agent/system-prompt.js";
 import { recordElapsed } from "./agent/profiling.js";
 
 import { initSkillSidecarDir } from "./skills/skill-sidecar.js";
@@ -31,8 +31,6 @@ import { buildActivationHeader } from "./skills/skill-manifest.js";
 import { TaskManager } from "./agent/task-manager.js";
 import { TaskStore } from "./agent/task-store.js";
 import { BgProcessManager } from "./agent/bg-process-manager.js";
-import { sessionFileRead, type SessionFsContext } from "./agent/session-fs.js";
-import { readTranscript, TranscriptReaderError } from "./agent/transcript-reader.js";
 
 import { isGitRepo, getRepoRoot, stripTokenFromGitUrl, detectPrimaryBranch } from "./skills/git.js";
 import { VerificationHarness } from "./agent/verification-harness.js";
@@ -82,18 +80,6 @@ import { BuiltinConfigProvider } from "./agent/builtin-config.js";
 import { ConfigCascade } from "./agent/config-cascade.js";
 
 import { initAssistantRegistry } from "./agent/assistant-registry.js";
-import {
-	deleteProposalFile,
-	editProposalFile,
-	isProposalType,
-	latestRev,
-	listProposalFiles,
-	parseProposalFile,
-	readProposalFile,
-	restoreSnapshot,
-	writeProposalFile,
-	type ProposalType,
-} from "./proposals/proposal-files.js";
 
 const execFileAsync = promisify(execFileCb);
 
@@ -1036,7 +1022,7 @@ async function handleApiRoute(
 	prStatusStore: PrStatusStore,
 	teamManager: TeamManager,
 	roleManager: RoleManager,
-	toolManager: ToolManager,
+	_toolManager: ToolManager,
 	projectContextManager: ProjectContextManager,
 	_bgProcessManager: BgProcessManager,
 	_staffManager: StaffManager,
@@ -2730,19 +2716,6 @@ async function handleApiRoute(
 		return;
 	}
 
-	// GET /api/sessions/:id/output — get final assistant output
-	const outputMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/output$/);
-	if (outputMatch && req.method === "GET") {
-		const id = outputMatch[1];
-		try {
-			const output = await sessionManager.getSessionOutput(id);
-			json({ output });
-		} catch {
-			json({ error: "Failed to get output" }, 500);
-		}
-		return;
-	}
-
 	// PATCH /api/sessions/:id — update session properties (title, colorIndex, etc.)
 	const patchMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)$/);
 	if (patchMatch && req.method === "PATCH") {
@@ -2861,412 +2834,6 @@ async function handleApiRoute(
 			}
 		}
 
-		json({ ok: true });
-		return;
-	}
-
-	// POST /api/sessions/:id/mark-read — record that the user viewed this session
-	const markReadMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/mark-read$/);
-	if (markReadMatch && req.method === "POST") {
-		const id = markReadMatch[1];
-		const ok = sessionManager.markSessionRead(id);
-		if (!ok) { json({ error: "session not found" }, 404); return; }
-		json({ ok: true });
-		return;
-	}
-
-	// ── Editable proposals (file-on-disk source of truth) ──────────────
-	// docs/design/editable-proposals.md §6.4
-	const proposalRouteMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/proposal\/([^/]+)(\/edit|\/seed|\/restore)?$/);
-	if (proposalRouteMatch) {
-		const sessionId = proposalRouteMatch[1];
-		const typeStr = proposalRouteMatch[2];
-		const suffix = proposalRouteMatch[3] || "";
-		if (!/^[A-Za-z0-9_-]+$/.test(sessionId)) {
-			json({ error: "Invalid sessionId" }, 400);
-			return;
-		}
-		if (!isProposalType(typeStr)) {
-			json({ error: `Unknown proposal type: ${typeStr}` }, 400);
-			return;
-		}
-		const proposalType = typeStr as ProposalType;
-		const proposalStateDir = bobbitStateDir();
-
-		// GET /api/sessions/:id/proposal/:type — read raw file
-		if (suffix === "" && req.method === "GET") {
-			try {
-				const content = await readProposalFile(proposalStateDir, sessionId, proposalType);
-				if (content === undefined) {
-					json({ ok: false, code: "FILE_NOT_FOUND", message: `No ${proposalType} proposal draft. Call propose_${proposalType} first.` }, 404);
-					return;
-				}
-				const contentType = proposalType === "goal" ? "text/markdown; charset=utf-8" : "application/yaml; charset=utf-8";
-				res.writeHead(200, { "Content-Type": contentType });
-				res.end(content);
-			} catch (err) {
-				json({ error: String((err as Error)?.message ?? err) }, 500);
-			}
-			return;
-		}
-
-		// DELETE /api/sessions/:id/proposal/:type
-		if (suffix === "" && req.method === "DELETE") {
-			try {
-				await deleteProposalFile(proposalStateDir, sessionId, proposalType);
-				if (_broadcastToSession) {
-					_broadcastToSession(sessionId, { type: "proposal_cleared", sessionId, proposalType });
-				}
-				res.writeHead(204);
-				res.end();
-			} catch (err) {
-				json({ error: String((err as Error)?.message ?? err) }, 500);
-			}
-			return;
-		}
-
-		// POST /api/sessions/:id/proposal/:type/edit — surgical edit
-		if (suffix === "/edit" && req.method === "POST") {
-			const body = await readBody(req);
-			if (!body || typeof body !== "object") {
-				json({ ok: false, code: "INVALID_BODY", message: "body must be JSON object" }, 400);
-				return;
-			}
-			const { old_text, new_text } = body as { old_text?: unknown; new_text?: unknown };
-			if (typeof old_text !== "string" || typeof new_text !== "string") {
-				json({ ok: false, code: "INVALID_BODY", message: "old_text and new_text must be strings" }, 400);
-				return;
-			}
-			try {
-				const result = await editProposalFile(proposalStateDir, sessionId, proposalType, old_text, new_text);
-				if (!result.ok) {
-					const status = result.code === "FILE_NOT_FOUND" ? 404 : 400;
-					json(result, status);
-					return;
-				}
-				if (_broadcastToSession) {
-					_broadcastToSession(sessionId, {
-						type: "proposal_update",
-						sessionId,
-						proposalType,
-						fields: result.parsed.fields,
-						rev: result.rev,
-						streaming: false,
-						source: "edit",
-					});
-				}
-				json({ ok: true, newContent: result.newContent, rev: result.rev });
-			} catch (err) {
-				json({ error: String((err as Error)?.message ?? err) }, 500);
-			}
-			return;
-		}
-
-		// POST /api/sessions/:id/proposal/:type/seed — called from propose_* execute()
-		if (suffix === "/seed" && req.method === "POST") {
-			const body = await readBody(req);
-			if (!body || typeof body !== "object") {
-				json({ ok: false, code: "INVALID_BODY", message: "body must be JSON object" }, 400);
-				return;
-			}
-			const args = (body as { args?: unknown }).args;
-			if (!args || typeof args !== "object" || Array.isArray(args)) {
-				json({ ok: false, code: "INVALID_BODY", message: "args must be an object" }, 400);
-				return;
-			}
-			try {
-				const writeRes = await writeProposalFile(proposalStateDir, sessionId, proposalType, args as Record<string, unknown>);
-				const parsed = await parseProposalFile(proposalStateDir, sessionId, proposalType);
-				if (!parsed.ok) {
-					json(parsed, 400);
-					return;
-				}
-				if (_broadcastToSession) {
-					_broadcastToSession(sessionId, {
-						type: "proposal_update",
-						sessionId,
-						proposalType,
-						fields: parsed.value.fields,
-						rev: writeRes.rev,
-						streaming: false,
-						source: "seed",
-					});
-				}
-				json({ ok: true, rev: writeRes.rev });
-			} catch (err) {
-				json({ error: String((err as Error)?.message ?? err) }, 500);
-			}
-			return;
-		}
-
-		// POST /api/sessions/:id/proposal/:type/restore — restore a snapshot
-		if (suffix === "/restore" && req.method === "POST") {
-			const body = await readBody(req);
-			if (!body || typeof body !== "object") {
-				json({ ok: false, code: "INVALID_BODY", message: "body must be JSON object" }, 400);
-				return;
-			}
-			const rev = (body as { rev?: unknown }).rev;
-			if (typeof rev !== "number" || !Number.isInteger(rev) || rev < 1) {
-				json({ ok: false, code: "INVALID_BODY", message: "rev must be a positive integer" }, 400);
-				return;
-			}
-			try {
-				const result = await restoreSnapshot(proposalStateDir, sessionId, proposalType, rev);
-				if (!result.ok) {
-					const status = (result as any).code === "SNAPSHOT_NOT_FOUND" ? 404 : 400;
-					json(result, status);
-					return;
-				}
-				if (_broadcastToSession) {
-					_broadcastToSession(sessionId, {
-						type: "proposal_update",
-						sessionId,
-						proposalType,
-						fields: result.fields,
-						rev: result.newRev,
-						streaming: false,
-						source: "restore",
-					});
-				}
-				json({ ok: true, newRev: result.newRev, fields: result.fields });
-			} catch (err) {
-				json({ error: String((err as Error)?.message ?? err) }, 500);
-			}
-			return;
-		}
-
-		json({ error: "Method not allowed" }, 405);
-		return;
-	}
-
-	// GET /api/sessions/:id/proposals — list all parsed proposal drafts for the session.
-	//
-	// Mirrors the WS-auth `proposal_update {source:"rehydrate"}` broadcast in
-	// `ws/handler.ts` but as a one-shot REST call. Used by the client's fast-path
-	// session switch-back (no fresh WS auth fires, so the broadcast doesn't run
-	// and the client's in-memory proposal slot would otherwise stay stale).
-	const proposalsListMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/proposals$/);
-	if (proposalsListMatch && req.method === "GET") {
-		const sessionId = proposalsListMatch[1];
-		if (!/^[A-Za-z0-9_-]+$/.test(sessionId)) {
-			json({ error: "Invalid sessionId" }, 400);
-			return;
-		}
-		const stateDir = bobbitStateDir();
-		try {
-			const types = await listProposalFiles(stateDir, sessionId);
-			const proposals: Array<{ proposalType: string; fields: Record<string, unknown>; rev: number }> = [];
-			for (const proposalType of types) {
-				const parsed = await parseProposalFile(stateDir, sessionId, proposalType);
-				if (parsed.ok) {
-					const rev = await latestRev(stateDir, sessionId, proposalType);
-					proposals.push({ proposalType, fields: parsed.value.fields, rev });
-				}
-			}
-			json({ proposals });
-		} catch (err) {
-			json({ error: String((err as Error)?.message ?? err) }, 500);
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/generate-title — auto-generate a title from chat history.
-	// Works for live sessions (calls SessionManager.autoGenerateTitle) and archived
-	// sessions (parses .jsonl). Used by the rename dialog when the session is not
-	// the currently focused one (no live WebSocket).
-	const genTitleMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/generate-title$/);
-	if (genTitleMatch && req.method === "POST") {
-		const id = genTitleMatch[1];
-		try {
-			const title = await sessionManager.generateTitleForAnySession(id);
-			if (!title) {
-				json({ error: "Could not generate title (session not found or no messages)" }, 404);
-				return;
-			}
-			json({ title });
-		} catch (err) {
-			json({ error: String((err as Error)?.message ?? err) }, 500);
-		}
-		return;
-	}
-
-	// PUT /api/sessions/:id/title — legacy rename endpoint
-	const titleMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/title$/);
-	if (titleMatch && req.method === "PUT") {
-		const id = titleMatch[1];
-		const body = await readBody(req);
-		const title = body?.title;
-		if (!title || typeof title !== "string") {
-			json({ error: "Missing title" }, 400);
-			return;
-		}
-		const ok = sessionManager.setTitle(id, title);
-		if (!ok) {
-			json({ error: "Session not found" }, 404);
-			return;
-		}
-		json({ ok: true });
-		return;
-	}
-
-	// GET /api/sessions/:id/tool-content/:messageIndex/:blockIndex — lazy-load full tool input content
-	const toolContentMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/tool-content\/(\d+)\/(\d+)$/);
-	if (toolContentMatch && req.method === "GET") {
-		const [, id, msgIdxStr, blkIdxStr] = toolContentMatch;
-		const messageIndex = parseInt(msgIdxStr, 10);
-		const blockIndex = parseInt(blkIdxStr, 10);
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		try {
-			const msgsResp = await session.rpcClient.getMessages();
-			const messages = msgsResp?.data?.messages || msgsResp?.data;
-			if (!Array.isArray(messages)) { json({ error: "Could not retrieve messages" }, 500); return; }
-			const msg = messages[messageIndex];
-			if (!msg) { json({ error: "Message not found" }, 404); return; }
-			const content = Array.isArray(msg.content) ? msg.content : [];
-			const block = content[blockIndex];
-			if (!block) { json({ error: "Block not found" }, 404); return; }
-			let toolContent = block.arguments?.content ?? block.input?.content;
-			// Fallback: text blocks (e.g. preview_open snapshot blocks in
-			// toolResult messages) store their payload in `block.text`.
-			if (toolContent === undefined && block.type === "text" && typeof block.text === "string") {
-				toolContent = block.text;
-			}
-			if (toolContent === undefined) { json({ error: "No content in block" }, 404); return; }
-			json({ content: toolContent });
-		} catch (err) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// GET /api/sessions/:id/transcript — paginated, regex-filterable transcript reader
-	// Backs the `read_session` tool extension. See `src/server/agent/transcript-reader.ts`.
-	const transcriptMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/transcript$/);
-	if (transcriptMatch && req.method === "GET") {
-		const [, targetId] = transcriptMatch;
-		// Resolve target session (live or persisted).
-		const targetPs = sessionManager.getPersistedSession(targetId);
-		if (!targetPs) { json({ error: "session_not_found" }, 404); return; }
-		if (!targetPs.agentSessionFile) { json({ error: "transcript_unavailable" }, 404); return; }
-
-		// Authorization: caller must belong to the same project as the target.
-		// Caller session id is propagated via `x-bobbit-session-id` header by the
-		// extension; if missing, fall back to allow (e.g. UI-initiated calls go
-		// through Bearer auth which already gates by project).
-		const callerSid = req.headers["x-bobbit-session-id"];
-		const callerSidStr = Array.isArray(callerSid) ? callerSid[0] : callerSid;
-		if (callerSidStr) {
-			const callerPs = sessionManager.getPersistedSession(callerSidStr);
-			if (callerPs && targetPs.projectId && callerPs.projectId && callerPs.projectId !== targetPs.projectId) {
-				json({ error: "permission_denied" }, 403); return;
-			}
-		}
-
-		// Parse query params.
-		const qp = url.searchParams;
-		function parseIntParam(name: string): number | undefined {
-			const raw = qp.get(name);
-			if (raw === null) return undefined;
-			const n = Number(raw);
-			if (!Number.isFinite(n)) {
-				throw new TranscriptReaderError("invalid_params", `${name} is not a number`);
-			}
-			return n;
-		}
-		try {
-			const params = {
-				offset: parseIntParam("offset"),
-				limit: parseIntParam("limit"),
-				pattern: qp.get("pattern") ?? undefined,
-				caseSensitive: qp.get("case_sensitive") === "1" || qp.get("case_sensitive") === "true",
-				context: parseIntParam("context"),
-				verbose: qp.get("verbose") === "1" || qp.get("verbose") === "true",
-			};
-			const ctx: SessionFsContext = { sandboxed: targetPs.sandboxed, projectId: targetPs.projectId };
-			const envelope = await readTranscript(params, {
-				readContent: () => sessionFileRead(ctx, targetPs.agentSessionFile, sandboxManager),
-			});
-			json(envelope);
-		} catch (err) {
-			if (err instanceof TranscriptReaderError) {
-				const status = err.code === "transcript_unavailable" ? 404 : 400;
-				json({ error: err.code, detail: err.message }, status);
-			} else {
-				jsonError(500, err, { error: "internal_error", detail: String(err) });
-			}
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/abort — force-abort a streaming session (graceful + force-kill)
-	const abortMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/abort$/);
-	if (abortMatch && req.method === "POST") {
-		const id = abortMatch[1];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		if (session.status !== "streaming") { json({ ok: true, status: session.status }); return; }
-		await sessionManager.forceAbort(id);
-		json({ ok: true, status: "idle" });
-		return;
-	}
-
-	// GET /api/sessions/:id/prompt-sections — return system prompt broken into labeled sections
-	const promptSectionsMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/prompt-sections$/);
-	if (promptSectionsMatch && req.method === "GET") {
-		const id = promptSectionsMatch[1];
-
-		// Try persisted snapshot first (captures the actual prompt at creation time)
-		const persisted = loadPersistedPromptSections(id);
-		if (persisted) {
-			json(persisted);
-			return;
-		}
-
-		// Fallback: reconstruct for legacy sessions without a persisted snapshot
-		const parts = sessionManager.getPromptParts(id);
-		if (!parts) { json({ error: "Session not found or no prompt data" }, 404); return; }
-
-		// Ensure tool docs are populated (they may have been injected at assemblePrompt time,
-		// but re-inject if missing to handle edge cases)
-		if (!parts.toolDocs && toolManager) {
-			parts.toolDocs = toolManager.getToolDocsForPrompt(parts.allowedTools, bobbitStateDir());
-		}
-
-		const sections = getPromptSections(parts);
-		const totalTokens = sections.reduce((sum, s) => sum + s.tokens, 0);
-		json({ sections, totalTokens });
-		return;
-	}
-
-	// GET /api/sessions/:id/draft?type=prompt — retrieve a draft
-	const draftGetMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/draft$/);
-	if (draftGetMatch && req.method === "GET") {
-		const id = draftGetMatch[1];
-		const type = url.searchParams.get("type");
-		if (!type) { json({ error: "Missing type query param" }, 400); return; }
-		const data = sessionManager.getDraft(id, type);
-		if (data === undefined) {
-			// Check if session exists at all
-			const session = sessionManager.getSession(id);
-			if (!session) { json({ error: "Session not found" }, 404); return; }
-			json({ error: "Draft not found" }, 404);
-			return;
-		}
-		json({ type, data });
-		return;
-	}
-
-	// DELETE /api/sessions/:id/draft?type=prompt — clear a draft
-	const draftDelMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/draft$/);
-	if (draftDelMatch && req.method === "DELETE") {
-		const id = draftDelMatch[1];
-		const type = url.searchParams.get("type");
-		if (!type) { json({ error: "Missing type query param" }, 400); return; }
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		sessionManager.deleteDraft(id, type);
 		json({ ok: true });
 		return;
 	}
